backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.5.0
Flask-SQLAlchemy==3.0.4
PyMySQL==1.0.3
reportlab==4.0.0
python-dotenv==1.0.0
